handle network error
