+++ src/canpeek/__main__.py
@@ -16,7 +16,6 @@ import json
 from typing import Dict, List, Optional, Any, TYPE_CHECKING
 from pathlib import Path
 from functools import partial
-import faulthandler
 import qdarktheme
 import inspect
 
@@ -69,7 +68,6 @@ from PySide6.QtCore import (
     QTimer,
     Signal,
     Qt,
-    QObject,
     QSortFilterProxyModel,
     QSettings,
 )
@@ -96,106 +94,16 @@ from .data_utils import (
     Connection,
 )
 
+from .can_utils import CANAsyncReader
+
 from .models import CANTraceModel, CANGroupedModel
 
 if TYPE_CHECKING:
     from __main__ import ProjectExplorer, CANBusObserver
 
 
-faulthandler.enable()
-
-
 # --- Data Structures ---
 TRACE_BUFFER_LIMIT = 5000
-
-
-# Replace the CANReaderThread class (around line 850)
-class CANAsyncReader(QObject):
-    frame_received = Signal(object)
-    error_occurred = Signal(str)
-    bus_state_changed = Signal(object)
-
-    def __init__(self, connection: Connection):
-        super().__init__()
-        self.connection = connection
-        self.running = False
-        self.bus = None
-        self.notifier = None
-        self.reader = None
-        self.read_task = None
-
-    async def start_reading(self):
-        """Start async CAN reading"""
-        try:
-            self.bus = can.Bus(
-                interface=self.connection.interface,
-                receive_own_messages=True,
-                **self.connection.config,
-            )
-
-            loop = asyncio.get_running_loop()
-            self.notifier = can.Notifier(self.bus, [self.rx_messages], loop=loop)
-            self.running = True
-            self.bus_state_changed.emit(self.bus.state)
-            return True
-
-        except Exception as e:
-            self.error_occurred.emit(
-                f"Failed to start CAN reading on {self.connection.name}: {e}"
-            )
-            return False
-
-    def rx_messages(self, msg: can.Message) -> None:
-        """Regular callback function. Can also be a coroutine."""
-        if msg.is_error_frame:
-            if self.bus:
-                self.bus_state_changed.emit(self.bus.state)
-            return
-
-        frame = CANFrame(
-            msg.timestamp,
-            msg.arbitration_id,
-            msg.data,
-            msg.dlc,
-            msg.is_extended_id,
-            msg.is_error_frame,
-            msg.is_remote_frame,
-            bus=self.connection.name,
-            connection_id=self.connection.id,
-            is_rx=msg.is_rx,
-        )
-        self.frame_received.emit(frame)
-
-    def stop_reading(self):
-        """Stop async CAN reading"""
-        self.running = False
-
-        if self.read_task and not self.read_task.done():
-            self.read_task.cancel()
-
-        if self.notifier:
-            self.notifier.stop()
-            self.notifier = None
-
-        if self.bus:
-            try:
-                self.bus.shutdown()
-            except Exception as e:
-                print(f"Error shutting down CAN bus on {self.connection.name}: {e}")
-            finally:
-                self.bus = None
-
-        self.reader = None
-        self.read_task = None
-
-    def send_frame(self, message: can.Message):
-        """Send a CAN frame"""
-        if self.bus and self.running:
-            try:
-                self.bus.send(message)
-            except Exception as e:
-                self.error_occurred.emit(f"Send error on {self.connection.name}: {e}")
-
 
 # --- UI Classes ---
 class DBCEditor(QWidget):
